# 🚀 Comprehensive Performance Optimizations