--- a/prisma/schema.prisma
+++ b/prisma/schema.prisma
@@ -38,6 +38,8 @@ model Product {
   priority        Int               @default(3)
   orderItems      OrderItem[]
   categories      ProductCategory[]
+
+  @@index([isActive, priority])
 }
 
 model Order {
@@ -53,6 +55,9 @@ model Order {
   createdAt     DateTime    @default(now())
   updatedAt     DateTime    @updatedAt
   orderItems    OrderItem[]
+
+  @@index([status, pickupDate])
+  @@index([customerName])
 }
 
 model OrderItem {
@@ -73,6 +78,8 @@ model OrderItem {
   productId        String
   order            Order   @relation(fields: [orderId], references: [id])
   product          Product @relation(fields: [productId], references: [id])
+
+  @@index([orderId, productId])
 }
 
 model UsageOption {
@@ -109,6 +116,7 @@ model ProductCategory {
   product    Product  @relation(fields: [productId], references: [id], onDelete: Cascade)
 
   @@unique([productId, categoryId])
+  @@index([categoryId, productId])
 }
 
 enum PriceType {